Make pb-jelly-gen Rust API optional
